+++ src/Autoload.php
@@ -23,6 +23,14 @@ class Autoload implements SectionInstanceInterface
     
     public $type;
     
+    /**
+     * Autoload Section
+     *
+     * @param ComposerReaderInterface $reader
+     * @param string $namespace The namespace like \\foo\\bar\\
+     * @param string $source The path which is root for the namespace like path/to
+     * @param string $type The type psr-4 or psr-0
+     */
     public function __construct(ComposerReaderInterface $reader, $namespace, $source, $type)
     {
         $this->reader = $reader;